M2-Step2b-2: cap_id_t 扩 int32_t (16→32 位,generation 8→24 bit)
核心扩位 (上次失败的点,现在根因已扫清):
- kernel_types.h: cap_id_t int16_t → int32_t, CAP_ID_T_MAX → INT32_MAX
- capability.c: CAP_GENERATION_MAX 0xFF → 0xFFFFFF (24 bit)
  SLOT_BITS 保持 7 (覆盖 128 全局 slot)
  cap 最大值 = (0xFFFFFF<<7)|0x7F = 0x7FFFFFFF = INT32_MAX (sign bit 永远 0)
- capability.h: cap_entry_t.generation uint16_t → uint32_t
- cap_encode/cap_next_generation 参数 uint16_t → uint32_t

根因修复 (Step 2a 已做):
1. cap_decode 删除 cap<0 守卫 → 高位合法 cap 不被误判
2. cap_encode_stays_positive 编译期断言保证合法 cap 永不为负

测试重构 (Step 2b-1 已做 test_driver/irq/service_model,本次补齐):
- test_syscall.c: 3 unpack + 3 pack 改 syscall_test_set_arg_pair
- test_allocator.c: 1 unpack + 1 pack 改 alloc_test_set_arg_pair
- test_fs_devfs.c 暂跳过 (VFS_ENABLE=0 不编译)

BSS 影响: cap_pool +256, TCB cap_set +4096, ep_cap_xfer_buffers +4096,
ep_server_reply_cap +4096, 共 +15KB (338K→353K,RP2350 SRAM 充足)。

影响面:
- ipc_cap_xfer_t 自动扩到 8 字节 (memcpy 用 sizeof,安全)
- 所有 side buffer 用类型维度,自动适应
- syscall 返回 (int)cap 路径: 合法 cap 永远 <= INT32_MAX,不与
  kern_err_t 负数命名空间冲突

测试: 编译通过 (text=428304, bss=353500),待硬件验证 2878/2878。
这是上次扩位失败的点,根因修复后重做。

+++ include/kernel/capability.h
@@ -54,7 +54,7 @@
  *============================================================================*/
 
 typedef struct {
-    uint16_t    generation; /* slot generation; prevents stale cap reuse */
+    uint32_t    generation; /* slot generation; prevents stale cap reuse (M2-Step2b: 16→32 位) */
     uint8_t     rights;     /* 权限位图 */
     uint8_t     owner;      /* 拥有者 task_id */
     void       *object;     /* 内核对象指针 */
